changed an 'int' to 'size_t,' to account for really big files

+++ include/lbann/data_readers/lbann_data_reader_imagenet_single.hpp
@@ -51,7 +51,7 @@ private:
   std::ifstream m_data_filestream;
   size_t m_file_size;
   std::vector<unsigned char> m_work_buffer;
-  std::vector<std::pair<int, int> > m_offsets; //stores: <offset, label>
+  std::vector<std::pair<size_t, int> > m_offsets; //stores: <offset, label>
   std::vector<unsigned char> m_pixels;
 
   void openDataStream();
